Fixed dropdown

+++ ui.v
@@ -123,8 +123,8 @@ pub:
 
 pub struct MouseMoveEvent {
 pub:
-	x            f64
-	y            f64
+	x            int
+	y            int
 	mouse_button int
 	// TODO enum
 }
